requirements.txt
openai==0.28
pillow==10.3.0
lxml==5.2.2
XlsxWriter==3.2.0
python-pptx==0.6.23
Flask==2.2.5
Flask-Cors==4.0.0
streamlit==1.34.0
